% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/registry.R
\name{load_registry}
\alias{load_registry}
\title{Load the registry of stored shortcuts into an environment.}
\usage{
load_registry(envir)
}
\arguments{
\item{envir}{environment. By default, the global environment.}
}
\description{
Load the registry of stored shortcuts into an environment.
}

